split mapping implementations into files per class
